% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{adpc}
\alias{adpc}
\title{Constructor for a (low dimensional) ADPROCLUS solution object}
\usage{
adpc(
  model,
  A,
  P,
  sse,
  totvar,
  explvar,
  iterations,
  timer,
  initialStart,
  C = NULL,
  B = NULL,
  runs = NULL
)
}
\arguments{
\item{model}{model}

\item{A}{A}

\item{P}{P}

\item{sse}{sse}

\item{totvar}{totvar}

\item{explvar}{explvar}

\item{iterations}{iterations}

\item{timer}{timer}

\item{initialStart}{initialstart}

\item{C}{C}

\item{B}{B}

\item{runs}{runs}
}
\value{
Object of class \code{adpc}
}
\description{
Yields an object of class \code{adpc}, which can be printed, plotted and summarized by the corresponding methods.
}
\examples{
#Create the information needed for an object of class adpc
x <- ADPROCLUS::CGdata[1:100,]
result <- adproclus(x, 3)
model <- result$model
A <- result$A
P <- result$P
totvar <- result$totvar
explvar <- result$explvar
iterations <- result$iterations
timer <- result$timer
initialStart <- result$ initialStart

#Create object of class adpc
result_object <- adpc(model, A, P, sse, totvar, explvar, iterations, timer, initialStart)

}
