% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\docType{class}
\name{adproclus-class}
\alias{adproclus-class}
\title{An S4 class to represent an ADPROCLUS solution.}
\description{
An S4 class to represent an ADPROCLUS solution.
}
\section{Slots}{

\describe{
\item{\code{clusters}}{A length-one numeric vector denoting the number of clusters in the solution}

\item{\code{Model}}{A model matrix M of the same size as the input data matrix.}

\item{\code{Membs}}{The n*k membership matrix of the clustering model}

\item{\code{Profs}}{The k*j profile matrix of the clustering model}

\item{\code{sse}}{A numeric denoting the residual sum of squares of the clustering model}

\item{\code{totvar}}{A numeric denoting the total sum of squares of the input data matrix}

\item{\code{explvar}}{A numeric denoting the proportion of variance in data that is accounted for by the clustering model.}

\item{\code{alg_iter}}{A numeric denoing the number of iterations of the ALS algorithm}

\item{\code{timer}}{A numeric denoting the elapsed time in seconds to obtain the clustering solution}

\item{\code{initialStart}}{A list containing information about the stat used to initialize the algorithm}
}}

