% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPROCLUS-package.R
\docType{package}
\name{ADPROCLUS-package}
\alias{ADPROCLUS}
\alias{ADPROCLUS-package}
\title{ADPROCLUS: Additive Profile Clustering Algorithms for R}
\description{
This package is an implementation of ADditive PROfile CLUStering (ADPROCLUS) in R. It can be used to obtain overlapping clustering models for object-by-variable data matrices.
}
\author{
\strong{Maintainer}: Rossbroich Julian \email{rossbroich@outlook.com}

Authors:
\itemize{
  \item Tom Wilderjans \email{t.f.wilderjans@fsw.leidenuniv.nl}
}

}
\keyword{internal}
