% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CGdata}
\alias{CGdata}
\title{Randomly generated data with underlying overlapping clusters.}
\format{
A data frame with 400 rows and 15 variables
}
\usage{
CGdata
}
\description{
A computer generated object-by-variable dataset with an underlying
nonrestricted overlapping clustering structure. For illustrative purposes
within the ADPROCLUS package only.
}
\keyword{datasets}
