% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getRational}
\alias{getRational}
\title{Generate initial rational start}
\usage{
getRational(data, centers)
}
\arguments{
\item{data}{Object-by-variable data matrix of class \code{matrix} or
\code{data.frame}.}

\item{centers}{either the number of clusters \emph{k}, or a matrix of initial
(distinct) cluster centres.}
}
\value{
\code{getRational} returns a list with the following components:
  \describe{
  \item{\code{type}}{A character string denoting the type of start
  ('Rational Start')}
  \item{\code{A}}{An initial Membership matrix}
  \item{\code{P}}{The corresponding initial Profile matrix} }
}
\description{
Generate an initial rational start for the Additive Profile Clustering
algorithm (see \code{\link{adproclus}}).
}
\details{
If \code{centers} is a number of clusters \emph{k}, an initial profile matrix
\strong{P} is generated by drawing \emph{k} randomly chosen, distinct, rows
from \code{data}. Alternatively, you can pass a user selected matrix of size
\emph{k} * \code{ncol(data)} with initial cluster profiles to the
\code{centers} argument. In both cases, a corresponding membership matrix
\strong{A} is subsequently estimated conditionally upon \strong{P} (for
details, see Depril et al., 2008, and Wilderjans et al., 2010).

\strong{Warning:} This function does \emph{not} obtain an ADPRCOLUS model. To
perform aditive profile clustering, see \code{\link{adproclus}}.
}
\examples{
getRational(ADPROCLUS::CGdata, 3)

}
\references{
Wilderjans, T. F., Ceulemans, E., Van Mechelen, I., & Depril, D.
  (2010). ADPROCLUS: a graphical user interface for fitting additive profile
  clustering models to object by variable data matrices. \emph{Behavior
  Research Methods, 43}(1), 56-65.

  Depril, D., Van Mechelen, I., & Mirkin, B. (2008). Algorithms for additive
  clustering of rectangular data tables. \emph{Computational Statistics and
  Data Analysis, 52,} 4923-4938.
}
\seealso{
\code{\link{getRandom}} for generating random starts and
  \code{\link{adproclus}} for details about membership and profile matrices.
}
