% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{name_clusters_adpc}
\alias{name_clusters_adpc}
\title{Naming the clusters of an ADPROCLUS solution}
\usage{
name_clusters_adpc(model, cluster_names)
}
\arguments{
\item{model}{ADPROCLUS solution (class: \code{adpc})}

\item{cluster_names}{Names of the clusters as list.}
}
\value{
The input model  of class \code{adpc} with adjusted cluster names.
}
\description{
Naming the clusters of an ADPROCLUS solution
}
\examples{
#example
}
