% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{name_components_adpc}
\alias{name_components_adpc}
\title{Naming the components of a low dimensional ADPROCLUS solution}
\usage{
name_components_adpc(model, component_names)
}
\arguments{
\item{model}{ADPROCLUS solution (class: \code{adpc})}

\item{component_names}{Names of the components of the low dimensional model as list.
Can be used to name variables for full dimensional model.}
}
\value{
The input model  of class \code{adpc} with adjusted component names.
}
\description{
Naming the components of a low dimensional ADPROCLUS solution
}
\examples{
#example
}
