% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_clusters_as_network}
\alias{plot_clusters_as_network}
\title{Plots a low dimensional ADPROCLUS solution as a network}
\usage{
plot_clusters_as_network(
  model,
  cluster_names = NULL,
  relative_overlap = TRUE,
  vertex_color = "antiquewhite",
  edge_color_low = "lavenderblush2",
  edge_color_high = "lavenderblush4"
)
}
\arguments{
\item{model}{the model to visualized}

\item{cluster_names}{optional: the names of the clusters}

\item{relative_overlap}{optional: choose whether you want to absolute number of observations on the edges}

\item{vertex_color}{optional: color of vertices}

\item{edge_color_low}{optional: lower end of color spectrum for edges}

\item{edge_color_high}{optional: upper end of color spectrum for edges}
}
\value{
nothing, outputs a plot
}
\description{
Plots a low dimensional ADPROCLUS solution as a network
}
\examples{
# Loading a test dataset into the global environment
x <- ADPROCLUS::CGdata

# Quick low dimensional clustering with K = 3 clusters and S = 1 dimensions
clust <- ldadproclus(x, 3, 1)

# Plot the overlapping the clusters
plot_clusters_as_network(clust)
}
