% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_scree_adpc_preselected}
\alias{plot_scree_adpc_preselected}
\title{Title}
\usage{
plot_scree_adpc_preselected(
  model_fit,
  title = "Scree plot of low dim ADPROCLUS pre-selected models",
  digits = 3
)
}
\arguments{
\item{digits}{}
}
\description{
Title
}
