% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotProfiles}
\alias{plotProfiles}
\title{Plot profile matrix of ADPROCLUS solution}
\usage{
plotProfiles(model, title = "Profiles of ADPROCLUS solution")
}
\arguments{
\item{model}{Object of class \code{adpc}. (Low dimensional) ADPROCLUS solution}

\item{title}{String. Default: "Profiles of ADPROCLUS solution"}
}
\value{
Invisibly returns the input model.
}
\description{
Plot profile matrix of ADPROCLUS solution
}
\examples{
#add exmple
}
