% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{print.adpc}
\alias{print.adpc}
\title{Print basic information on ADPROCLUS solution}
\usage{
\method{print}{adpc}(x, title = "ADPROCLUS solution", ...)
}
\arguments{
\item{x}{ADPROCLUS solution (class: \code{adpc})}

\item{title}{String. Default: "ADPROCLUS solution"}

\item{...}{ignored}
}
\description{
For an object of class adpc as input, this method prints basic information
about the ADPROCLUS solution represented by the object. Works for both full and low dimensional solutions.
}
\examples{
x <- ADPROCLUS::CGdata[1:100,]
model <- adproclus(x, 3)
print(model)
}
