% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{print.summary.adpc}
\alias{print.summary.adpc}
\title{Printing (low dim) ADPROCLUS summary of class \code{summary.adpc}}
\usage{
\method{print}{summary.adpc}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary.adpc}}

\item{...}{ignored}
}
\value{
\code{summary.adpc} object invisibly
}
\description{
Prints a set of summary statistics for a (low dimensional) ADPROCLUS solution as provided by the function \code{summary.adpc()}.
A number of parameters for how the results should be printed can be passed as an argument to \code{summary.adpc()}
which then passes it on to this method.
}
\examples{
x <- ADPROCLUS::CGdata[1:100,]
model <- adproclus(x, 3)
summary(model)
}
