% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{select_by_CHull}
\alias{select_by_CHull}
\title{Title
only works when the unexplained variance/SSE is decreasing in the number of components
if for a certain number of clusters exactly 2 models where estimated, this procedure will always choose the one with the better model fit}
\usage{
select_by_CHull(model_fit, percentage_fit = 1e-04)
}
\arguments{
\item{PercentageFit}{}
}
\description{
Title
only works when the unexplained variance/SSE is decreasing in the number of components
if for a certain number of clusters exactly 2 models where estimated, this procedure will always choose the one with the better model fit
}
