% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADPROCLUS-package.R
\docType{package}
\name{ADPROCLUS-package}
\alias{ADPROCLUS}
\alias{ADPROCLUS-package}
\title{ADPROCLUS: Additive Profile Clustering Algorithms for R}
\description{
Obtain overlapping clustering models for object-by-variable data matrices using the additive profile clustering (ADPROCLUS) method. Also contains the low dimensional ADPROCLUS method, for simultaneous dimension reduction and overlapping clustering.
}
\author{
\strong{Maintainer}: Henry Heppe \email{heppe.henry@gmail.com}

Authors:
\itemize{
  \item Julian Rossbroich \email{rossbroich@outlook.com}
  \item Jeffrey Durieux
  \item Tom Wilderjans \email{t.f.wilderjans@fsw.leidenuniv.nl}
}

}
\keyword{internal}
