% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus-package.R
\docType{package}
\name{adproclus-package}
\alias{adproclus-package}
\title{adproclus: Additive Profile Clustering Algorithms}
\description{
Obtain overlapping clustering models for object-by-variable data matrices using the Additive Profile Clustering (ADPROCLUS) method. Also contains the low dimensional ADPROCLUS method for simultaneous dimension reduction and overlapping clustering. For reference see Depril, Van Mechelen, Mirkin (2008) \doi{10.1016/j.csda.2008.04.014} and Depril, Van Mechelen, Wilderjans (2012) \doi{10.1007/s00357-012-9112-5}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/henry-heppe/adproclus}
  \item Report bugs at \url{https://github.com/henry-heppe/adproclus/issues}
}

}
\author{
\strong{Maintainer}: Henry Heppe \email{heppe.henry@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Julian Rossbroich
  \item Jeffrey Durieux
  \item Tom Wilderjans
}

}
\keyword{internal}
