% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{adproclusLD}
\alias{adproclusLD}
\title{Low dimensional ADPROCLUS}
\usage{
adproclusLD(
  data,
  nclusters,
  ncomponents,
  start_allocation = NULL,
  nrandomstart = 1,
  nsemirandomstart = 1,
  SaveAllStarts = FALSE
)
}
\arguments{
\item{data}{Object-by-variable data matrix of class \code{matrix} or
\code{data.frame}.}

\item{nclusters}{Number of clusters to be used. Must be a positive integer.}

\item{ncomponents}{Number of components (dimensions) to which the profiles should be restricted. Must be a positive integer.}

\item{start_allocation}{Matrix of binary values as starting allocation for first run. Default is \code{NULL}.}

\item{nrandomstart}{Number of random starts (see \link{getRandom}). Can be zero.}

\item{nsemirandomstart}{Number of semi-random starts (see \link{getRational})). Can be zero.}

\item{SaveAllStarts}{logical. If \code{TRUE}, the results of all algorithm
starts are returned. By default, only the best solution is retained.}
}
\value{
a list of relevant information about the resulting model
}
\description{
Low dimensional ADPROCLUS
}
\examples{
# Loading a test dataset into the global environment
x <- ADPROCLUS::CGdata

# Low dimensional clustering with K = 3 clusters
# where the resulting profiles can be characterized in S = 1 dimensions (components)
clust <- adproclusLD(x, 3, 1)
}
