% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{cluster_means}
\alias{cluster_means}
\title{Cluster Means based on Original Variables}
\usage{
cluster_means(data, model, digits = 3)
}
\arguments{
\item{data}{Object-by-variable matrix. Can contain other variables than
the ADPROCLUS model.}

\item{model}{ADPROCLUS solution (class: \code{adpc}). Low dimensional model
possible.}

\item{digits}{Integer. The number of decimal places that all decimal numbers will
be rounded to.}
}
\value{
Cluster-by-variable dataframe where the values are the cluster means
for the given variable.
}
\description{
Obtain a cluster-by-variable dataframe where the values are the cluster means
for the given variable. Takes as input a (low dimensional) ADPROCLUS model
of class \code{adpc} and the dataset the model was estimated on.
The dataset does not actually have to have the same variables as the one
the model was trained on, since this function can also be used to automatically
compute the cluster means for a different set of variables for the same
set of observations. The last row \code{Cl0} is the baseline cluster consisting
of all the observations that were not assigned to a cluster,
if this cluster is not empty.
}
\details{
This can be e.g. useful if one has a patient by symptom dataset, but only
clusters on a subset of the symptoms. In this case this function nonetheless
return the cluster means for all symptoms to describe the patients in a
cluster more closely.

The output of this function is different from the last output matrix in the
\code{summary()} method. The former is in terms of the original variables
while the latter is in terms of the approximated model variables.
}
\examples{
# Obtain data, compute model, report cluster means
x <- CGdata
model <- adproclus(x, 3)
cluster_means(data = x, model = model)
}
