% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ldadproclus}
\alias{ldadproclus}
\title{Low dimensional ADPROCLUS}
\usage{
ldadproclus(
  data,
  nclusters,
  ncomponents,
  start_allocation = NULL,
  nrandomstart = 1,
  nsemirandomstart = 1,
  SaveAllStarts = FALSE
)
}
\arguments{
\item{data}{the data}

\item{nclusters}{the number of clusters}

\item{ncomponents}{the number of dimensions (the number of components)}

\item{start_allocation}{a possible start allocation}

\item{nrandomstart}{the number of random starts}

\item{SaveAllStarts}{option to save all starts}

\item{randomstart}{type of random start}
}
\value{
a list of relevant information about the resulting model
}
\description{
Low dimensional ADPROCLUS
}
\examples{
some example
}
