% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{plot.adpc}
\alias{plot.adpc}
\title{Plotting a (low dimensional) ADPROCLUS solution}
\usage{
\method{plot}{adpc}(x, type = "Network", title = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{adpc}. (Low dimensional) ADPROCLUS solution}

\item{...}{ignored}
}
\value{
invisibly returns the argument
}
\description{
Plotting a (low dimensional) ADPROCLUS solution
}
\examples{
#add example
}
