% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plot_scree_adpc}
\alias{plot_scree_adpc}
\title{Title}
\usage{
plot_scree_adpc(
  model_list,
  title = "Scree plot of ADPROCLUS models",
  unexplvar = TRUE,
  digits = 2
)
}
\arguments{
\item{digits}{}
}
\description{
Title
}
