% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{plotVarsByComp}
\alias{plotVarsByComp}
\title{Plot variable to component matrix of ADPROCLUS solution}
\usage{
plotVarsByComp(model, title = "B' of Low Dimensional ADPROCLUS Solution")
}
\arguments{
\item{model}{Object of class \code{adpc}. Must be \strong{Low dimensional} ADPROCLUS solution}
}
\value{
Invisibly returns the input model.
}
\description{
Plot variable to component matrix of ADPROCLUS solution
}
\examples{
#add examples
}
