% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adproclus_classes.R
\name{summary.adpc}
\alias{summary.adpc}
\title{Summarize important information on ADPROCLUS solution}
\usage{
\method{summary}{adpc}(
  object,
  title = "ADPROCLUS solution",
  digits = 3,
  matrix_rows = 10,
  matrix_cols = 15,
  ...
)
}
\arguments{
\item{object}{ADPROCLUS solution (class: \code{adpc})}

\item{...}{ignored}
}
\value{
invisibly returns object of class \code{summary.adpc}
}
\description{
Summarize important information on ADPROCLUS solution
}
\examples{
x <- ADPROCLUS::CGdata[1:100,]
model <- adproclus(x, 3)
summary(model)
}
